HealthAssistantBackend/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
attrs==23.2.0
blinker==1.8.2
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
cymem==2.0.8
dataclasses-json==0.6.6
decorator==5.1.1
distro==1.9.0
dnspython==2.6.1
elevenlabs==0.2.26
executing==2.0.1
Flask==3.0.3
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
ipython==8.24.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.2.0
langchain-community==0.2.0
langchain-core==0.2.0
langchain-text-splitters==0.2.0
langcodes==3.4.0
langsmith==0.1.59
language_data==1.2.0
marisa-trie==1.1.1
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib-inline==0.1.7
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
numpy==1.26.4
openai==0.28
orjson==3.10.3
packaging==23.2
pandas==2.2.2
parso==0.8.4
pinecone==4.0.0
preshed==3.0.9
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
pymongo==4.7.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
scikit-learn==1.4.2
scipy==1.13.0
setuptools==69.5.1
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
spacey==0.1.1
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.30
srsly==2.4.8
stack-data==0.6.3
tenacity==8.3.0
thinc==8.2.3
threadpoolctl==3.5.0
tqdm==4.66.4
traitlets==5.14.3
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
websockets==12.0
Werkzeug==3.0.3
yarl==1.9.4
transformers
torch
accelerate
groq
python-dotenv
scikit-learn
